+++ webpage/static/dashboard/style.css
@@ -66,7 +66,7 @@ table {
   
   /* Add a background color on hover */
   .btn-grp button:hover {
-    background-color: #3e8e41;
+    background-color: #bdb6b2;
   }
 
   .bold{
