+++ webpage/templates/dashboard.html
@@ -4,20 +4,20 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Dashboard</title>
-    <link rel="stylesheet" href="../util/utils.css">
-    <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="/static/dashboard/style.css">
+    <link rel="stylesheet" href="/static/utils/utils.css">
 </head>
 <body>
     <header class="container">
         <nav class="flex">
             <div class="left flex items-center">
-                <img src="../img/redhen.png" alt="">
+                <img src="/static/utils/redhen.png" alt="">
                 <ul class="flex items-center justify-center">
                     <li>Red Hen Anonymizer</li>
                 </ul>
             </div>
             <div class="right flex">
-                <img src="../img/user.png" alt="">
+                <img src="/static/utils/user.png" alt="">
             </div>
         
         </nav>
@@ -26,7 +26,7 @@
     <div class="flex setup">
         <div class="flex left-body">
             <!-- add image -->
-            <img src="../img/user.png" alt="Avatar"> 
+            <img src="/static/utils/user.png" alt="Avatar"> 
             
         <div class="info">
             <span id ="name" class="bold ">
@@ -38,7 +38,10 @@
             </span>
             
         </div>
-
+        <div class="btn-grp">
+            <button onclick="location.href='main'" type="button">
+                + Create new experiment</button>
+        </div>
         </div>
         
         <div class="flex right-body">
